Slightly resized the platform
I found that the length of the platform was too big, and so I made it shorter. I was a little stupid and only edited it inside the main level map itself, but as long as we copy paste that platform to make new ones it will have the same effect.